+++ inst/examples/example-indexSelfConstruction.R
@@ -1,0 +1,13 @@
+# using distance measures
+indexSelfConstruction(boeker, 1, 2, c(3:11), method = "euclidean")
+indexSelfConstruction(boeker, 1, 2, c(3:11), method = "manhattan")
+indexSelfConstruction(boeker, 1, 2, c(3:11), method = "minkowski", p = 3)
+
+# using correlation measures
+indexSelfConstruction(boeker, 1, 2, c(3:11), method = "pearson")
+indexSelfConstruction(boeker, 1, 2, c(3:11), method = "spearman")
+
+# printing the results (biplot only works with)
+cp <- indexSelfConstruction(boeker, 1, 2, c(3:11))
+cp$grid  # grid with self, ideal and others
+biplot2d(cp$grid, center = 4)  # midopoint centering